refactor(frontend): extract PhotoSection widget from listing_form.dart
- Move _PhotoSection (now public PhotoSection) and _PhotoThumb (stays
  private) into widgets/photo_section.dart
- ListingForm imports and uses PhotoSection from the new file
- No duplicate code between files

<em>🤖 Co-created with Claude</em>
